requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.6.15
charset-normalizer==2.0.12
Django==4.0.5
django-admin-interface==0.20.0
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-countries==7.3.2
django-extensions==3.2.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
graphviz==0.20.1
idna==3.3
Pillow==9.2.0
pycodestyle==2.9.1
pydot==1.4.2
PyJWT==2.4.0
pyparsing==3.0.9
pytz==2022.1
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.9
